Capstone Week 3 Part 3
I used the geo coordinates obtained from the geocoding API in part 2, to build requests to send to foursquare's API and retrieve nearby venues. I group the postal codes by the sum of the venue categories and run it through a kMeans clustering model to classify the zip codes into 5 different categories. I plot them on a folium map by color in the end